Add Xaya as new coin with basic chain handling (#889)
* Split out deserialisation of CAuxPow.

The current DeserializerAuxPow has code for reading a block header
including (potentially) an auxpow.  In this change, we split out the
logic for reading the actual CAuxPow instance to its own function.

This makes the code easier to read, as the different steps are more
clearly separated.  It will also be useful for working with CAuxPow
instances that are part of a different serialisation format (e.g.
for the Xaya blockchain).

* Add Xaya coin with deserialiser.

This defines a new coin for Xaya (https://github.com/xaya/xaya), with
the appropriate deserialiser methods.  Xaya is based on Namecoin, but
uses its own special twist on auxpow for the PoW, so we need to handle
that properly in block-header deserialisation.

Xaya also has names like Namecoin (but with slightly changed semantics),
but handling for them (i.e. building a name index in addition to the
address index) will be added later on.

* Keep more data in truncated Xaya headers.

When Xaya headers are truncated (because they are checkpointed), we still
need to send the algorithm and bits with them.  That is needed for the
difficulty retargeting of Xaya to work correctly.

* Define peers for Xaya.

We will run ElectrumX instances on the seeder VPSes.  Set those in PEERS
for mainnet (there are no peers for testnet).